Moved CalendarUtils.kt to core module
